feat: oxide infrastructure provider (#2)
* feat: oxide infrastructure provider

Built the Oxide infrastructure provider for Sidero Omni so that users
can automatically deploy Oxide instances for Omni to provision and
configure as Kubernetes nodes. This is a
[dynamic infrastructure provider](https://docs.siderolabs.com/omni/infrastructure-and-extensions/infrastructure-providers#dynamic-providers)
that will automatically create and destroy Oxide instances as needed.

This is still a work in progress. There's more to learn about how
infrastructure providers work and are meant to be implemented.

* misc: get provider working

* misc: cli and env var configuration

* misc: prepare for demo

* misc: initial provider refactor