provider: use machine information during deprovision (#4)
Updated `github.com/siderolabs/omni/client` to include the change in
https://github.com/siderolabs/omni/commit/a91eabdf1e6ae3161485a8426b648b58489e53cb.

This ensures that the machine information is no longer `nil` during
calls to `Deprovision`. The `Deprovision` method was also updated to
make use of this machine information instead of trying to unmarshal the
machine class from raw YAML. That means we can now deprovision the Oxide
instance by its ID rather than needing to fetch the project and instance
name, which is consistent with how other infrastructure providers work.