fix: prevent duplicate provider function calls
Updated `github.com/siderolabs/omni/client` to include the change in
https://github.com/siderolabs/omni/commit/c2cbf34b02e1288cf7d55bc0ed444513fa912d18.

This ensures both `ProvisionSteps` and `Deprovision` are no longer
called twice, allowing us to eliminate idempotency logic in the provider
should we choose to do so. It also is less confusing to users since they
will no longer see logs showing that a function was called right after
the function completed successfully.